requirements.txt
pangea-sdk
ecommClover
Flask
PyCryptodome
